@@ -9,15 +9,18 @@
     "dev": "nodemon src/index.ts"
   },
   "dependencies": {
-    "express": "^4.17.1",
-    "axios": "^0.21.1",
+    "express": "^4.18.2",
+    "http-proxy-middleware": "^2.0.6",
     "cors": "^2.8.5",
-    "dotenv": "^8.2.0"
+    "dotenv": "^16.3.1"
   },
   "devDependencies": {
-    "typescript": "^4.1.2",
-    "ts-node": "^9.1.1",
-    "nodemon": "^2.0.7"
+    "typescript": "^5.3.3",
+    "ts-node": "^10.9.2",
+    "nodemon": "^3.0.2",
+    "@types/express": "^4.17.21",
+    "@types/cors": "^2.8.17",
+    "@types/node": "^20.10.4"
   },
   "repository": {
     "type": "git",
