@@ -9,22 +9,26 @@
     "test": "jest"
   },
   "dependencies": {
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
+    "express": "^4.18.2",
+    "jsonwebtoken": "^9.0.2",
     "bcryptjs": "^2.4.3",
-    "mongoose": "^5.10.9",
-    "dotenv": "^8.2.0"
+    "mongoose": "^8.0.3",
+    "cors": "^2.8.5",
+    "dotenv": "^16.3.1"
   },
   "devDependencies": {
-    "typescript": "^4.1.2",
-    "ts-node": "^9.0.0",
-    "jest": "^26.6.3",
-    "@types/jest": "^26.0.0",
-    "@types/node": "^14.14.31",
-    "@types/express": "^4.17.8"
+    "typescript": "^5.3.3",
+    "ts-node": "^10.9.2",
+    "jest": "^29.7.0",
+    "@types/jest": "^29.5.8",
+    "@types/node": "^20.10.4",
+    "@types/express": "^4.17.21",
+    "@types/cors": "^2.8.17",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/jsonwebtoken": "^9.0.5"
   },
   "engines": {
-    "node": ">=12.0.0"
+    "node": ">=18.0.0"
   },
   "license": "MIT"
 }
