@@ -2,24 +2,31 @@
   "name": "word-processor-backend",
   "version": "1.0.0",
   "description": "Backend for the Word Processing application",
-  "main": "src/app.ts",
+  "main": "dist/app.js",
   "scripts": {
-    "start": "ts-node src/app.ts",
+    "start": "node dist/app.js",
+    "dev": "ts-node-dev --respawn --transpile-only src/app.ts",
     "build": "tsc",
-    "dev": "ts-node-dev src/app.ts"
+    "clean": "rimraf dist"
   },
   "dependencies": {
-    "express": "^4.17.1",
-    "body-parser": "^1.19.0",
-    "cors": "^2.8.5"
+    "express": "^4.18.2",
+    "cors": "^2.8.5",
+    "helmet": "^7.1.0",
+    "express-rate-limit": "^7.1.5",
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
-    "typescript": "^4.0.0",
-    "ts-node": "^9.0.0",
-    "ts-node-dev": "^1.0.0"
+    "typescript": "^5.3.3",
+    "ts-node": "^10.9.1",
+    "ts-node-dev": "^2.0.0",
+    "@types/express": "^4.17.21",
+    "@types/cors": "^2.8.17",
+    "@types/uuid": "^9.0.7",
+    "rimraf": "^5.0.5"
   },
   "engines": {
-    "node": ">=12.0.0"
+    "node": ">=18.0.0"
   },
   "license": "MIT"
 }
