@@ -1,43 +1,36 @@
 {
   "name": "media-gallery-platform",
-  "version": "1.0.0",
-  "description": "An image and video gallery application.",
-  "main": "src/server/index.js",
+  "version": "2.0.0",
+  "description": "Modern media gallery with React 18 and TypeScript",
   "scripts": {
-    "start": "node src/server/index.js",
-    "dev": "nodemon src/server/index.js",
-    "client": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev:server": "tsx watch src/server/index.ts",
+    "dev:client": "cd src/client && npm run dev",
+    "build:server": "tsc -p src/server/tsconfig.json",
+    "build:client": "cd src/client && npm run build",
+    "start:server": "node dist/server/index.js",
+    "start": "npm run start:server",
+    "dev": "concurrently \"npm run dev:server\" \"npm run dev:client\""
   },
   "dependencies": {
-    "express": "^4.18.2",
+    "express": "^4.19.2",
     "cors": "^2.8.5",
-    "mongoose": "^8.0.3",
+    "mongoose": "^8.3.0",
     "multer": "^1.4.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
-    "axios": "^1.6.2"
+    "helmet": "^7.1.0",
+    "express-rate-limit": "^7.2.0",
+    "joi": "^17.12.3",
+    "dotenv": "^16.4.5"
   },
   "devDependencies": {
-    "nodemon": "^3.0.2",
-    "typescript": "^5.3.3",
+    "typescript": "^5.4.5",
+    "tsx": "^4.7.2",
+    "concurrently": "^8.2.2",
     "@types/express": "^4.17.21",
     "@types/cors": "^2.8.17",
     "@types/multer": "^1.4.11",
-    "@types/react": "^18.2.45",
-    "@types/react-dom": "^18.2.18",
-    "@types/node": "^20.10.5"
+    "@types/node": "^20.12.7"
   },
-  "keywords": [
-    "gallery",
-    "images",
-    "videos",
-    "react",
-    "express"
-  ],
-  "author": "Your Name",
+  "keywords": ["gallery", "media", "react", "typescript", "modern-ui"],
+  "author": "Media Gallery Team",
   "license": "MIT"
 }
