@@ -1,24 +1,33 @@
 import express from 'express';
 import mongoose from 'mongoose';
-import bodyParser from 'body-parser';
+import cors from 'cors';
+import dotenv from 'dotenv';
 import userRoutes from './routes/userRoutes';
 
+dotenv.config();
+
 const app = express();
-const PORT = process.env.PORT || 5000;
+const PORT = process.env.PORT || 3000;
 
 // Middleware
-app.use(bodyParser.json());
+app.use(cors());
+app.use(express.json());
+
+// Mongoose configuration for v8.x
+mongoose.set('strictQuery', false);
 
 // Database connection
-mongoose.connect(process.env.MONGODB_URI || 'mongodb://localhost:27017/auth-service', {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-})
+mongoose.connect(process.env.MONGODB_URI || 'mongodb://mongodb:27017/auth-service')
 .then(() => console.log('MongoDB connected'))
 .catch(err => console.error('MongoDB connection error:', err));
 
 // Routes
-app.use('/api/auth', userRoutes);
+app.use('/auth', userRoutes);
+
+// Health check
+app.get('/health', (req, res) => {
+    res.json({ status: 'Auth service is running' });
+});
 
 // Start server
 app.listen(PORT, () => {
