@@ -1,18 +1,24 @@
 import express from 'express';
 import mongoose from 'mongoose';
-import bodyParser from 'body-parser';
+import cors from 'cors';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 const app = express();
 const PORT = process.env.PORT || 3000;
 
 // Middleware
-app.use(bodyParser.json());
+app.use(cors());
+app.use(express.json());
+
+// Mongoose configuration
+mongoose.set('strictQuery', false);
 
 // MongoDB connection
-mongoose.connect(process.env.MONGODB_URI || 'mongodb://localhost:27017/project-service', {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-});
+mongoose.connect(process.env.MONGODB_URI || 'mongodb://mongodb:27017/project-service')
+.then(() => console.log('MongoDB connected'))
+.catch(err => console.error('MongoDB connection error:', err));
 
 // Define Project Schema
 const projectSchema = new mongoose.Schema({
@@ -26,14 +32,27 @@ const Project = mongoose.model('Project', projectSchema);
 
 // Routes
 app.get('/projects', async (req, res) => {
-    const projects = await Project.find();
-    res.json(projects);
+    try {
+        const projects = await Project.find();
+        res.json(projects);
+    } catch (error) {
+        res.status(500).json({ error: error.message });
+    }
 });
 
 app.post('/projects', async (req, res) => {
-    const newProject = new Project(req.body);
-    await newProject.save();
-    res.status(201).json(newProject);
+    try {
+        const newProject = new Project(req.body);
+        await newProject.save();
+        res.status(201).json(newProject);
+    } catch (error) {
+        res.status(400).json({ error: error.message });
+    }
+});
+
+// Health check
+app.get('/health', (req, res) => {
+    res.json({ status: 'Project service is running' });
 });
 
 // Start the server
