@@ -1,18 +1,24 @@
 import express from 'express';
 import mongoose from 'mongoose';
-import bodyParser from 'body-parser';
+import cors from 'cors';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 const app = express();
-const PORT = process.env.PORT || 3001;
+const PORT = process.env.PORT || 3000;
 
 // Middleware
-app.use(bodyParser.json());
+app.use(cors());
+app.use(express.json());
+
+// Mongoose configuration
+mongoose.set('strictQuery', false);
 
 // MongoDB connection
-mongoose.connect(process.env.MONGODB_URI || 'mongodb://localhost:27017/taskdb', {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-});
+mongoose.connect(process.env.MONGODB_URI || 'mongodb://mongodb:27017/task-service')
+.then(() => console.log('MongoDB connected'))
+.catch(err => console.error('MongoDB connection error:', err));
 
 // Task schema
 const taskSchema = new mongoose.Schema({
@@ -25,24 +31,45 @@ const Task = mongoose.model('Task', taskSchema);
 
 // Routes
 app.get('/tasks', async (req, res) => {
-    const tasks = await Task.find();
-    res.json(tasks);
+    try {
+        const tasks = await Task.find();
+        res.json(tasks);
+    } catch (error) {
+        res.status(500).json({ error: error.message });
+    }
 });
 
 app.post('/tasks', async (req, res) => {
-    const newTask = new Task(req.body);
-    await newTask.save();
-    res.status(201).json(newTask);
+    try {
+        const newTask = new Task(req.body);
+        await newTask.save();
+        res.status(201).json(newTask);
+    } catch (error) {
+        res.status(400).json({ error: error.message });
+    }
 });
 
 app.put('/tasks/:id', async (req, res) => {
-    const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true });
-    res.json(updatedTask);
+    try {
+        const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true });
+        res.json(updatedTask);
+    } catch (error) {
+        res.status(400).json({ error: error.message });
+    }
 });
 
 app.delete('/tasks/:id', async (req, res) => {
-    await Task.findByIdAndDelete(req.params.id);
-    res.status(204).send();
+    try {
+        await Task.findByIdAndDelete(req.params.id);
+        res.status(204).send();
+    } catch (error) {
+        res.status(400).json({ error: error.message });
+    }
+});
+
+// Health check
+app.get('/health', (req, res) => {
+    res.json({ status: 'Task service is running' });
 });
 
 // Start server
